feat: use lanyard
