+++ src/pages/index.astro
@@ -4,7 +4,7 @@ import Intro from "../components/Intro.astro";
 import Projects from "../components/Projects.astro";
 import WaveDivider from "../components/WaveDivider.astro";
 import Webring from "../components/Webring.astro";
-import SpotifyNowPlaying from "../components/SpotifyNowPlaying.astro";
+import LanyardActivity from "../components/LanyardActivity.astro";
 import Socials from "../components/Socials.astro";
 ---
 
@@ -16,7 +16,7 @@ import Socials from "../components/Socials.astro";
     <Projects />
     <WaveDivider color="flamingo" speed={100} size={0.8} direction="right" />
 
-    <SpotifyNowPlaying />
+    <LanyardActivity />
     <WaveDivider color="mauve" speed={100} size={0.8} direction="left" />
 
     <Webring />
